requirements.txt
pydantic>=2.0.0
PyYAML>=6.0.1
aiohttp>=3.9.1
click>=8.1.7
rich>=13.7.0
aiosqlite>=0.19.0
setuptools>=69.0.3
aiosmtplib>=3.0.1

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0
pylint>=3.0.3
